@@ -4,113 +4,152 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Validation\Rules\Password;
+use Illuminate\Support\Facades\Validator;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class AuthController extends Controller
 {
-    public function login(Request $request)
+    /**
+     * Register a new user
+     */
+    public function register(Request $request): JsonResponse
     {
-        $request->validate([
-            'email' => 'required|email',
-            'password' => 'required',
-            'device_name' => 'required',
-        ]);
-
-        $user = User::where('email', $request->email)->first();
-
-        if (!$user || !Hash::check($request->password, $user->password)) {
-            return response()->json([
-                'message' => 'The provided credentials are incorrect.'
-            ], 401);
-        }
-
-        if (!$user->is_active) {
-            return response()->json([
-                'message' => 'Your account is inactive.'
-            ], 403);
-        }
-
-        $token = $user->createToken($request->device_name)->plainTextToken;
-
-        return response()->json([
-            'user' => $user,
-            'token' => $token,
-        ]);
-    }
-
-    public function register(Request $request)
-    {
-        $validated = $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
-            'password' => ['required', 'confirmed', Password::defaults()],
-            'phone' => 'nullable|string|max:20',
-            'device_name' => 'required|string',
+            'password' => 'required|string|min:8|confirmed',
+            'phone_number' => 'nullable|string|max:20',
         ]);
 
-        $user = User::create([
-            'name' => $validated['name'],
-            'email' => $validated['email'],
-            'password' => $validated['password'],
-            'phone' => $validated['phone'] ?? null,
-            'role' => 'collector', // Default role
-        ]);
-
-        $token = $user->createToken($validated['device_name'])->plainTextToken;
-
-        return response()->json([
-            'user' => $user,
-            'token' => $token,
-        ], 201);
-    }
-
-    public function logout(Request $request)
-    {
-        $request->user()->currentAccessToken()->delete();
-
-        return response()->json([
-            'message' => 'Logged out successfully'
-        ]);
-    }
-
-    public function user(Request $request)
-    {
-        return response()->json($request->user());
-    }
-
-    public function updateProfile(Request $request)
-    {
-        $validated = $request->validate([
-            'name' => 'sometimes|string|max:255',
-            'phone' => 'nullable|string|max:20',
-        ]);
-
-        $request->user()->update($validated);
-
-        return response()->json($request->user());
-    }
-
-    public function updatePassword(Request $request)
-    {
-        $request->validate([
-            'current_password' => 'required',
-            'password' => ['required', 'confirmed', Password::defaults()],
-        ]);
-
-        if (!Hash::check($request->current_password, $request->user()->password)) {
+        if ($validator->fails()) {
             return response()->json([
-                'message' => 'Current password is incorrect.'
+                'success' => false,
+                'errors' => $validator->errors()
             ], 422);
         }
 
-        $request->user()->update([
-            'password' => $request->password,
+        $user = User::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => Hash::make($request->password),
+            'phone_number' => $request->phone_number,
+            'is_active' => true,
         ]);
 
+        // Assign default role (e.g., 'collector')
+        // $user->roles()->attach(Role::where('slug', 'collector')->first());
+
+        $token = JWTAuth::fromUser($user);
+
         return response()->json([
-            'message' => 'Password updated successfully'
+            'success' => true,
+            'message' => 'User registered successfully',
+            'user' => $user,
+            'token' => $token,
+            'token_type' => 'bearer',
+        ], 201);
+    }
+
+    /**
+     * Login user
+     */
+    public function login(Request $request): JsonResponse
+    {
+        $validator = Validator::make($request->all(), [
+            'email' => 'required|email',
+            'password' => 'required|string',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'errors' => $validator->errors()
+            ], 422);
+        }
+
+        $credentials = $request->only('email', 'password');
+
+        if (!$token = auth('api')->attempt($credentials)) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Invalid credentials'
+            ], 401);
+        }
+
+        $user = auth('api')->user();
+        
+        if (!$user->is_active) {
+            auth('api')->logout();
+            return response()->json([
+                'success' => false,
+                'message' => 'User account is inactive'
+            ], 403);
+        }
+
+        // Update last login
+        $user->update(['last_login_at' => now()]);
+
+        return $this->respondWithToken($token, $user);
+    }
+
+    /**
+     * Get authenticated user
+     */
+    public function me(): JsonResponse
+    {
+        $user = auth('api')->user();
+        
+        return response()->json([
+            'success' => true,
+            'user' => $user->load('roles.permissions')
+        ]);
+    }
+
+    /**
+     * Logout user
+     */
+    public function logout(): JsonResponse
+    {
+        auth('api')->logout();
+
+        return response()->json([
+            'success' => true,
+            'message' => 'Successfully logged out'
+        ]);
+    }
+
+    /**
+     * Refresh token
+     */
+    public function refresh(): JsonResponse
+    {
+        try {
+            $token = auth('api')->refresh();
+            $user = auth('api')->user();
+
+            return $this->respondWithToken($token, $user);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Token refresh failed'
+            ], 401);
+        }
+    }
+
+    /**
+     * Helper to format token response
+     */
+    protected function respondWithToken($token, $user): JsonResponse
+    {
+        return response()->json([
+            'success' => true,
+            'user' => $user,
+            'token' => $token,
+            'token_type' => 'bearer',
+            'expires_in' => auth('api')->factory()->getTTL() * 60
         ]);
     }
 }
