@@ -3,9 +3,5 @@
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
-    return response()->json([
-        'message' => 'TransacTrack API',
-        'version' => '1.0.0',
-        'status' => 'operational',
-    ]);
+    return view('welcome');
 });
